Fix: Address 4 Integer Type Issues seen on Arm32 Launchpad
There were 4 issues (1 error, 3 warnings) that showed up in the
Launchpad Arm-32 build following merge of the Windows-support PR, all
to do with integer sign and width conversions. Addressing those issues
here.

+++ src/common/http.cc
@@ -1223,7 +1223,7 @@ namespace Pistache::Http
 
         PST_OUT(writeHeaders(writer.headers(), *buf));
 
-        const size_t len = sb.st_size;
+        const size_t len = static_cast<size_t>(sb.st_size);
 
         PST_OUT(writeHeader<Header::ContentLength>(os, len));
